include: move per-cpu storage helpers into main helper directory
