+++ scheds/rust/scx_wd40/src/bpf/lb_domain.bpf.c
@@ -16,7 +16,6 @@
 #include "intf.h"
 #include "types.h"
 #include "lb_domain.h"
-#include "percpu.h"
 
 #include <errno.h>
 #include <stdbool.h>
